factoried the site security and check_resources using check_security.
