+++ nvflare/ha/overseer/overseer.py
@@ -97,4 +97,4 @@ def refresh():
 
 
 if __name__ == "__main__":
-    app.run(host="localhost", port=6000)
+    app.run()
