+++ nvflare/private/fed/server/site_security.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import logging
-from typing import List
+from typing import List, Tuple
 
 from nvflare.apis.event_type import EventType
 from nvflare.apis.fl_constant import FLContextKey
@@ -29,7 +29,7 @@ class SiteSecurityFilter(CommandFilter):
     def __init__(self) -> None:
         super().__init__()
 
-    def pre_command(self, conn: Connection, args: List[str]) -> (bool, str):
+    def pre_command(self, conn: Connection, args: List[str]) -> Tuple[bool, str]:
         engine = conn.app_ctx
         command = args[0]
 
@@ -48,15 +48,6 @@ class SiteSecurityFilter(CommandFilter):
             with engine.new_context() as fl_ctx:
                 fl_ctx.set_prop(FLContextKey.COMMAND_NAME, command, sticky=False)
                 engine.fire_event(EventType.SECURITY_CHECK, fl_ctx)
-
-                authentication_result = fl_ctx.get_prop(FLContextKey.AUTHENTICATION_RESULT, True)
-                if not authentication_result:
-                    reasons = fl_ctx.get_prop(FLContextKey.AUTHENTICATION_REASON, {})
-                    messages = self._get_messages(reasons)
-                    logger.error(f"Authentication failed. Reason: {messages}")
-                    fl_ctx.remove_prop(FLContextKey.AUTHENTICATION_RESULT)
-                    fl_ctx.remove_prop(FLContextKey.AUTHENTICATION_REASON)
-                    filter_succeed = False
 
                 authorization_result = fl_ctx.get_prop(FLContextKey.AUTHORIZATION_RESULT, True)
                 if not authorization_result:
